Experimental version of double-buffered collector - x2 speedup in collection
- instead of attempting to perform atomic operations of locking the active buffer on every call, we simply store the call in the small buffer, that on filling up is sent via circular buffer to the analyzer thread. This significantly simplified calls_collector_thread::track() call, which led to the decrease of total overhead from ~125 to ~65 cycles per profiled call;
- polyq library added to support circular buffers for feeding /consuming of the first-level buffers.

+++ frontend/src/untested/tables_ui.cpp
@@ -77,10 +77,10 @@ namespace micro_profiler
 		_statistics_lv->set_columns_model(_columns_main);
 		_children_lv->set_columns_model(_columns_children);
 
-		_connections.push_back(_threads_cb->selection_changed += [model] (unsigned int index) {
+		_connections.push_back(_threads_cb->selection_changed += [model] (size_t index) {
 			unsigned id;
 
-			if (model->get_threads()->get_key(id, index))
+			if (model->get_threads()->get_key(id, static_cast<unsigned>(index)))
 				model->set_filter([id] (const functions_list::value_type &v) { return id == v.first.second;	});
 			else
 				model->set_filter();
